refactor: split the registration page into components
